Open combine pipe before full
Useful in combination with AutoDrive.
Open combine pipe even before full if there is an AutoDrive unloader assigned to the combine and the combine fill level is over the precall level, and Dynamic Chase Distance is activated in AutoDrive settings. On the AIVehicleExtension mod simply set "wait during unloading" to off so the active unloading happens.
Update to latest AD version to be able to use this feature. Doesn't give any log error if AD is not installed or updated.